% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_varclass.R
\name{convert_varclass}
\alias{convert_varclass}
\title{convert_varclass}
\usage{
convert_varclass(df)
}
\arguments{
\item{df}{dataframe of genomic alterations}
}
\value{
data frame with variant classes converted
}
\description{
if conversion of variant classes is selected, this function will do that
}
\details{
variant classes are converted based on this schema:\cr
splicing -> Splicing\cr
stopgain or stoploss -> Nonsense\cr
nonsynonymous -> Missense\cr
frameshift_deletion or insertion -> Frameshift\cr
nonframeshift_deletion or insertion -> InFrame\cr
upstream - Promoter\cr
IntragenicDeletion -> Frameshift\cr
}
\examples{
TODO
}

