% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.matrix.R
\name{paste.matrix}
\alias{paste.matrix}
\title{paste.matrix}
\usage{
paste.matrix(M1, M2)
}
\arguments{
\item{M1}{matrix of genomic alterations}

\item{M2}{matrix of genomic alterations}
}
\value{
merged matrix of M1 and M2
}
\description{
takes two genomic evetn matrices as input and merges them. the matrices do not have the have the same dimensions. If a gene and a sample have two different alterations, they will be merged seperated with a comma
}
\examples{
TODO
}

