% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSort.R
\name{memoSort}
\alias{memoSort}
\title{sampleSort}
\usage{
memoSort(M, geneOrder = geneOrder, annotations = NULL,
  annotation_order = NULL)
}
\arguments{
\item{M}{matrix of alterations.}

\item{annotations}{this holds the per sample annotation values. the matrix will be split into n sub matrices and sorted then merged at the end.}

\item{annotation_order}{(Required) specify the order of the annotation. If not present, R will sort it alphabetically}

\item{geneName}{If certain genes needs to be on the top, this value should provide them in a list}
}
\value{
matrix with samples and genes sorted
}
\description{
sampleSort
}
\examples{
TODO2
}

