% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oncoPrint.R
\name{oncoPrint}
\alias{oncoPrint}
\title{oncoPrint}
\usage{
oncoPrint(df, sort = TRUE, convert = TRUE, total_samples = NA,
  geneName = NA, annotation = NA, annotation_order = NA,
  merge_scnas = F, df2 = NA, alteration_score = list(Amplification = 3,
  Deletion = 2, Nonsense = 2.8, Frameshift = 2.5, Splicing = 2.5, InFrame = 2,
  Promoter = 2, Mutation = 1, Missense = 1, Present = 1, NotTested = 0, None =
  0), printSamples = T)
}
\arguments{
\item{df}{this is the dataframe used to plot oncoprints. required colnames: Sample, Gene, VarClass}

\item{sort}{Boelean indicating whether genes should be sorted or now (default: True)}

\item{convert}{Boelean indicating whether varclasses should be converted to more standard names (default: True)}

\item{total_samples}{Total number of samples. If not given, total row numbers of input will be used}

\item{geneName}{IF given, this gene sits on top of the oncoprint, it won't be used in sorting}

\item{annotation_order}{a list of annotation items to be displayed}

\item{merge_scnas}{Bolean indicating whether to plot scnas and mutations together superimposed (default: False)}

\item{alteration_score}{this list determines the relative importance of different genomic alterations. Amplification > Deletion > Mutations etc.}

\item{printSamples}{Bolean indicating whether sample names should be printed under the oncoprint}

\item{annotations}{If given, these will be used for pre-clustering}

\item{second_df}{if merge_scnas is True, this is the second set of events to be merged}
}
\value{
oncoPrint image
}
\description{
oncoPrint
}
\examples{
TODO
}

