% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoPrint.R
\name{oncoPrint}
\alias{oncoPrint}
\title{oncoPrint}
\usage{
oncoPrint(df, sort = TRUE, convert = TRUE, total_samples = NULL,
  geneName = NULL, annotation = NULL, annotation_order = NULL,
  merge_scnas = F, df2 = NULL, onco_colors = list(Mutation = "#26A818",
  Missense = "#26A818", Nonsense = "black", Splicing = "#A05E35", Frameshift =
  "#A05E35", Promoter = "#2986E2", InFrame = "#F26529", Present = "darkorchid2",
  NotPresent = "#DCD9D3", NotTested = "darkgrey", del = "red", LOH =
  "darkkhaki", homodel = "brown4", CNLOH = "deepskyblue", Amplification =
  "#EA2E49", Deletion = "#174D9D", Yes = "#155B6B", No = "#12C8F9", Unknown =
  "azure1", Fusion = "#D38C1F"), alteration_score = list(Amplification = 5,
  Fusion = 4.5, Deletion = 4, Nonsense = 2.8, Frameshift = 2.5, Splicing = 2.5,
  InFrame = 2, Promoter = 2, Mutation = 1, Missense = 1, Present = 1, NotTested
  = 0, None = 0, NotPresent = 0, Yes = 0, No = 0, del = 3, homodel = 2, LOH =
  1.5, CNLOH = 1), printSamples = T, xpadding = 0.1, ypadding = 0.1)
}
\arguments{
\item{df}{this is the dataframe used to plot oncoprints. required colnames: Sample, Gene, VarClass}

\item{sort}{Boelean indicating whether genes should be sorted or now (default: True)}

\item{convert}{Boelean indicating whether varclasses should be converted to more standard names (default: True)}

\item{total_samples}{Total number of samples. If not given, total row numbers of input will be used}

\item{geneName}{If given, genes are not automatically sorted. Instead this variable is used for sorting. Partial list of genes can be input}

\item{annotation_order}{This is the order of annotation categories to display. Required if annotations are given}

\item{merge_scnas}{Bolean indicating whether to plot scnas and mutations together superimposed (default: False)}

\item{df2}{If multiple classes of alterations or clinical data points are to be merged, input your list of dataframes here.}

\item{alteration_score}{this list determines the relative importance of different genomic alterations. Amplification > Deletion > Mutations etc.}

\item{printSamples}{Bolean indicating whether sample names should be printed under the oncoprint}

\item{xpadding}{numerical value of the padding between two consecutive data columns}

\item{ypadding}{numerical value of the padding between two consecutive data rows}

\item{annotations}{If given, these will be used for creating sub-groups of samples. Clustering will happen within annotatoion groups and then merged}
}
\value{
oncoPrint image and a list of values
}
\description{
This function creates onco prints of genomic data based on mutations, copy number alterations, fusions, and other user defined categories. For each category 3 fields are required : Sample ID, Gene, and Variant Class. There are options to display binary events as well as continous variables.
}
\examples{
TODO
}

